toolbox/ColossalAI/v0.4.4/patches/requirements/requirements.txt
numpy
tqdm
psutil
packaging
pre-commit
rich
click
fabric
contexttimer
ninja
# torch>=2.2.0,<=2.4.0
safetensors
einops
pydantic
ray
sentencepiece
google
protobuf
transformers>=4.39.3 # BI适配的版本太落后，直接使用官方的版本
peft>=0.7.1
# bitsandbytes>=0.39.0
rpyc==6.0.0
fastapi
uvicorn==0.29.0
galore_torch
# diffusers==0.29.0
